# Fix floating point tests in Rust implementation 