--- a/rust/src/consts.rs
+++ b/rust/src/consts.rs
@@ -47,3 +47,25 @@ pub const GRID_ROWS: usize = 5;
 
 // Minimum length of a code that can be shortened.
 pub const MIN_TRIMMABLE_CODE_LEN: usize = 6;
+
+// What to multiple latitude degrees by to get an integer value. There are three pairs representing
+// decimal digits, and five digits in the grid.
+pub const LAT_INTEGER_MULTIPLIER: i64 = (ENCODING_BASE
+    * ENCODING_BASE
+    * ENCODING_BASE
+    * GRID_ROWS
+    * GRID_ROWS
+    * GRID_ROWS
+    * GRID_ROWS
+    * GRID_ROWS) as i64;
+
+// What to multiple longitude degrees by to get an integer value. There are three pairs representing
+// decimal digits, and five digits in the grid.
+pub const LNG_INTEGER_MULTIPLIER: i64 = (ENCODING_BASE
+    * ENCODING_BASE
+    * ENCODING_BASE
+    * GRID_COLUMNS
+    * GRID_COLUMNS
+    * GRID_COLUMNS
+    * GRID_COLUMNS
+    * GRID_COLUMNS) as i64;