--- a/rust/src/private.rs
+++ b/rust/src/private.rs
@@ -1,6 +1,6 @@
 use consts::{
-    CODE_ALPHABET, ENCODING_BASE, GRID_COLUMNS, GRID_ROWS, LATITUDE_MAX, LONGITUDE_MAX,
-    NARROW_REGION_PRECISION, PAIR_CODE_LENGTH,
+    CODE_ALPHABET, ENCODING_BASE, GRID_ROWS, LATITUDE_MAX, LONGITUDE_MAX,
+    PAIR_CODE_LENGTH,
 };
 
 use interface::encode;
@@ -31,9 +31,9 @@ pub fn clip_latitude(latitude_degrees: f64) -> f64 {
 
 pub fn compute_latitude_precision(code_length: usize) -> f64 {
     if code_length <= PAIR_CODE_LENGTH {
-        return ENCODING_BASE.powf((code_length as f64 / -2f64 + 2f64).floor());
+        return (ENCODING_BASE as f64).powf((code_length as f64 / -2f64 + 2f64).floor());
     }
-    ENCODING_BASE.powi(-3i32) / GRID_ROWS.powf(code_length as f64 - PAIR_CODE_LENGTH as f64)
+    (ENCODING_BASE as f64).powf(-3f64) / GRID_ROWS.pow((code_length - PAIR_CODE_LENGTH) as u32) as f64
 }
 
 pub fn prefix_by_reference(pt: Point<f64>, code_length: usize) -> String {
@@ -47,47 +47,4 @@ pub fn prefix_by_reference(pt: Point<f64>, code_length: usize) -> String {
     );
     code.drain(code_length..);
     code
-}
-
-// Apply "gravity" towards closest integer value, if current value is closer than given threshold.
-// This is a way to compensate aggregated error caused by floating point precision restriction.
-fn near(value: f64, error: f64) -> f64 {
-    let target = (value + error).trunc();
-    if value.trunc() != target {
-        target
-    } else {
-        value
-    }
-}
-
-pub fn narrow_region(digit: usize, lat: &mut f64, lng: &mut f64) {
-    if digit == 0 {
-        *lat /= ENCODING_BASE;
-        *lng /= ENCODING_BASE;
-    } else if digit < PAIR_CODE_LENGTH {
-        *lat *= ENCODING_BASE;
-        *lng *= ENCODING_BASE;
-    } else {
-        *lat *= GRID_ROWS;
-        *lng *= GRID_COLUMNS
-    }
-    *lat = near(*lat, NARROW_REGION_PRECISION);
-    *lng = near(*lng, NARROW_REGION_PRECISION);
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn near_applied() {
-        let value = 3.0f64 - NARROW_REGION_PRECISION * 2.;
-        assert_eq!(near(value, NARROW_REGION_PRECISION), value);
-    }
-
-    #[test]
-    fn near_not_applied() {
-        let value = 3.0f64 - NARROW_REGION_PRECISION;
-        assert_eq!(near(value, NARROW_REGION_PRECISION), 3.0f64);
-    }
 }