Fix final floating point tests in rust implementation by using integer operations for encode and decode.