@drinckes 2019-05-28
Impact:
* encoding is slower (~1usec to ~3usec)
* decoding is faster (~25usec to ~14usec)