@drinckes 2019-05-28
Hi Sergey

There was one failing test - I ended up changing the implementation in this library to the integer based solution that e.g. the Java library uses, and just translated it to rust.

Given that I don't really know rust at all, could you take a look and see if there's anything terrible I've done?